Pare down the exported symbol list.
We don't want to pollute the global namespace with function names like
"newq" but I don't want to have to name every non-static function in a
library Smear might use "CONFLICT_AVOIDANCE_newq". This change links
libsmear.a then makes all the non-exported symbols internal.

This adds a dependency for GCC or things that support GCC's
"-fvisibility" command line option as well as its
"__attribute__((visibility))". To the extent that other compilers
support this functionality (and most do) but spell it differently,
this should not be hard to port. Compilers that don't support it may
just yield a library with a more polluted namespace. That's fine with
me for now.