Add SRT_wait_for_empty, change SRT_wait_for_idle's semantics.
After the cancellation changes, wait_for_idle effectively changed
meanings. Since timed events are now mixed in with all the rest of the
events in the queue, waiting for the queue to be empty meant stalling
until all the timers had fired.

In updating the tutorial examples, it became clear that this was not
good behavior for a function that might be called in a
mainloop. However, since it's still useful, I added SRT_wait_for_empty
to preserve the old (post-cancellation) semantics.

Also updated the tutorial to live in the new reality.