Add basic 128-bit unsigned multiple-precision arithmetic.
Comparison, addition, subtraction, and multiplication are all simple,
expedient algorithms that are overly-specified to 128-bits.

However, the division algorithm (the difficult one) is general purpose.
It is from Per Brinch Hanson's paper "Multiple-Length Division Revisited,"
as explained in the comment in number.c above div128.  There are a few
shortcuts there that make assumptions about the size of the operands, but
those would be pretty to abstract if we want to extend to larger naturals.