Use QueryPerformanceCounter as the source of time on Windows.
Per the MSDN article "Acquiring high-resolution time stamp," QPC is the
proper source of accurate, precise, large, monotonic time that reflects
actual time passage (e.g., correctly accounting for standby/hibernate).
Moreover, it works as we need it to on all versions of Windows back to and
including WinXP, and works indentically for 32-bit and 64-bit Windows.

The disadvantages are that it cannot be meaningfully synchronized to an
external time reference, and that on some older processors it can be more
expensive to query than other methods, taking as long as 1us, which is
driven by hardware.